backend/requirements.txt
flask[async]
openai
python-dotenv
flask_cors
supabase
clerk_backend_api
asyncio